backend/requirements.txt
Django==3.2
djangorestframework==3.12.0
djoser==0.4.1
django-cors-headers==3.7.0
mysqlclient==2.0.3
Pillow==8.2.0
djangorestframework-simplejwt==4.6.0
django-filter==2.4.0

